wip: pull util out
