+++ mozilla_bitbar_devicepool/main.py
@@ -10,11 +10,10 @@ import zipfile
 from mozilla_bitbar_devicepool import (
     TESTDROID,
     configuration,
-    download_file,
     logger,
     modulepath,
 )
-
+from mozilla_bitbar_devicepool.util.network import download_file
 from mozilla_bitbar_devicepool.bitbar.runs import run_test_for_project
 from mozilla_bitbar_devicepool.test_run_manager import TestRunManager
